requirements.txt
joblib
tensorflow
numpy
scikit-learn
gdown
requests
fastapi
python-multipart
Pillow
python-multipart
geopy
roboflow
uvicorn
